feat: user profile avatar, 20-08-2024

+++ src/pages/api/[[...user]].ts
@@ -1,5 +1,6 @@
 import { deleteData, retrieveData, updateData } from "@/lib/firebase/service";
 import type { NextApiRequest, NextApiResponse } from "next";
+import jwt from "jsonwebtoken";
 
 export default async function handler(
   req: NextApiRequest,
@@ -13,22 +14,43 @@ export default async function handler(
     });
     res.status(200).json({ status: true, statusCode: 200, data });
   } else if (req.method === "PUT") {
-    const { id, data } = req.body;
-    await updateData("users", id, data, (result: boolean) => {
-      if (result) {
-        res.status(200).json({ status: true, statusCode: 200 });
-      } else {
-        res.status(400).json({ status: false, statusCode: 400 });
+    const { user }: any = req.query;
+    const { data } = req.body;
+    const token = req.headers.authorization?.split(" ")[1] || "";
+    jwt.verify(
+      token,
+      process.env.NEXTAUTH_SECRET || "",
+      async (err: any, decoded: any) => {
+        if (decoded && decoded.role === "admin") {
+          await updateData("users", user[1], data, (result: boolean) => {
+            if (result) {
+              res.status(200).json({ status: true, statusCode: 200 });
+            } else {
+              res.status(400).json({ status: false, statusCode: 400 });
+            }
+          });
+        }
       }
-    });
+    );
   } else if (req.method === "DELETE") {
     const { user }: any = req.query;
-    await deleteData("users", user[1], (result: boolean) => {
-      if (result) {
-        res.status(200).json({ status: true, statusCode: 200 });
-      } else {
-        res.status(400).json({ status: false, statusCode: 400 });
+    const token = req.headers.authorization?.split(" ")[1] || "";
+    jwt.verify(
+      token,
+      process.env.NEXTAUTH_SECRET || "",
+      async (err: any, decoded: any) => {
+        if (decoded && decoded.role === "admin") {
+          await deleteData("users", user[1], (result: boolean) => {
+            if (result) {
+              res.status(200).json({ status: true, statusCode: 200 });
+            } else {
+              res.status(400).json({ status: false, statusCode: 400 });
+            }
+          });
+        } else {
+          res.status(403).json({ status: false, statusCode: 200 });
+        }
       }
-    });
+    );
   }
 }
